services/detection-engine/requirements.txt
kafka-python==2.0.2
opensearch-py==2.3.1
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
pyyaml==6.0.1
requests==2.31.0
redis==4.6.0
prometheus-client==0.17.1